==> Notebook/2_Preprocessing_Coretax.ipynb <==
# 🔄 Preprocessing Data: Cleaning & Preparation 🛠️  
Setelah berhasil melakukan **scraping** data dari berita, langkah selanjutnya adalah **preprocessing**.  
Tahap ini bertujuan untuk membersihkan dan menyiapkan data agar siap untuk analisis lebih lanjut.  

---

## 🔧 **Libraries Used**

import pandas as pd
import re
from Sastrawi.StopWordRemover.StopWordRemoverFactory import StopWordRemoverFactory
from Sastrawi.Stemmer.StemmerFactory import StemmerFactory
from collections import Counter

## 📥 **Import Data CSV**

# Import data
file_path = "../data/hasil_scrap_berita.xlsx"
df = pd.read_excel(file_path)
df

#Hanya akan menampilkan kolom berikut
df.loc[:,["SUMBER","PENERBIT","JUDUL", "content", "tags", "date"]]

## 📌 **Tahapan Preprocessing**

### ✅ **1. Handling Missing Values** → Menghapus atau mengisi data yang kosong  

print(df.isnull().sum())

Data tidak ada yang kosong

### ✅ **2. Lowercasing** → Mengubah teks menjadi huruf kecil untuk konsistensi

# Ubah ke lowercase hanya pada kolom teks
df['JUDUL'] = df['JUDUL'].str.lower()
df['tags'] = df['tags'].str.lower()
df['content'] = df['content'].str.lower()

df.head()

### ✅ **3. Special Characters Removal** → menghapus kata spesial pada berita

def clean_text(text):
    # text = re.sub(r'^[a-zA-Z\s,]+[-\s–]+', '', text)
    text = re.sub(r'\([^)]*\)', '', text) #menghapus teks dalam kurung
    text = re.sub(r'http[s]?://\S+', '', text) #menghapus url
    text = re.sub(r'^[^—-]+[-—]\s*', '', text) ## Menghapus kata penerbit, kota di awal
    text = re.sub(r'^\d{1,2}\s\w+\s\d{4}\s*[-:]\s*', '', text) #menghapus tanggal
    text = re.sub(r'&\w+;', ' ', text)       # karakter html, Contoh: &nbsp; &quot;
    text = re.sub(r'&#\d+;', ' ', text)      # karalter html, Contoh: &#x27;
    text = re.sub(r'\d+', '', text)          # Menghapus angka
    return text

# Terapkan pada kolom 'content'
df['clean_text'] = df['content'].apply(clean_text)


df.loc[:,["content", "clean_text"]]

### ✅ **4. Punctuations removal** → Menghapus tanda baca dan simbol yang tidak relevan.

def clean_special_chars(text):
    # Hilangkan emoji dan simbol unicode lainnya
    text = re.sub(r'[^\x00-\x7F]+', ' ', text)  
    
    # Hilangkan tanda baca dan simbol matematika
    text = re.sub(r'[^\w\s]', ' ', text)

    # Hilangkan extra whitespace
    text = re.sub(r'\s+', ' ', text).strip()
    
    return text

# Terapkan
df['content_clean'] = df['clean_text'].apply(clean_special_chars)


df.loc[:,["clean_text", "content_clean"]]

### ✅ **5. Tokenization** → Memisahkan teks menjadi kata-kata atau unit kecil lainnya.

def tokenize(text):
    return re.findall(r'\b\w+\b', text.lower())

# Apply
df['tokens'] = df['content_clean'].apply(tokenize)

# Cek hasil tokenisasi
df[["content_clean", "tokens"]].head()

### ✅ **6. Stopword Removal** → Menghapus kata-kata umum yang tidak bermakna dalam analisis

stop_factory = StopWordRemoverFactory()
stop_words = set(stop_factory.get_stop_words())

filtered_stop_words = stop_words.copy()
new_stopwords = [
    "ada", "lalu", "guna", "klik", "sini", "beri", "bagai", "atas",
    "bulan", "sama", "kini", "menjadi", "tersebut", "baca", "lebih",
    "klik", "guna", "iklan", "sebut", "jadi", "paja",
    "berita", "data", "disini", "langganan", "akun",
    "direktorat", "jenderal", "administrasi", "masa", "tahun",
    "negara", "pemerintah", "menteri", "keuangan", "ditjen",
    "indonesia", "pengguna""iklan","detik", "hari", "nomor", "bacan",
    "tahunan", "artikel", "with", "pajak", "coretax"
]

def remove_stopwords(tokens):
    cleaned_tokens = [
        token.strip() for token in tokens  # Hapus spasi atau karakter aneh
        if len(token) > 3  # Hapus kata tidak penting
    ]
    # Menghapus stopwords dari token
    return [t for t in cleaned_tokens if t.lower() not in filtered_stop_words]

# Terapkan ke DataFrame
df['stopword_removal'] = df['tokens'].apply(remove_stopwords)

# Cek hasil 
df[["tokens", "stopword_removal"]].head()

### ✅ **7.  Stemming** → Mengubah kata menjadi bentuk dasarnya

# Membuat objek stemmer
stem_factory = StemmerFactory()
stemmer = stem_factory.create_stemmer()

# Fungsi untuk melakukan stemming pada tokens
def stem_tokens(tokens):
    return [stemmer.stem(token) for token in tokens]

# Terapkan stemming ke DataFrame
df['stemming_output'] = df['stopword_removal'].apply(stem_tokens)

# Cek hasil 
df[["stopword_removal", "stemming_output"]].head()

### ✅ **8. Rare Words Removal** → Menghapus kata yang jarang muncul.

# Hanya ambil kata dari kolom stopword_removal
all_words_stopword = [word for tokens in df["stopword_removal"] for word in tokens]

# Hitung frekuensi spesifik hanya untuk kolom stopword_removal
word_counts_stopword = Counter(all_words_stopword)

# Tentukan rare words berdasarkan kolom ini saja
rare_threshold = 5
rare_words_stopword = {word for word, count in word_counts_stopword.items() if count < rare_threshold}

# Bersihkan
def remove_rare_stopword(tokens):
    return [word for word in tokens if word not in rare_words_stopword]

df['stopword_removal_clean'] = df['stopword_removal'].apply(remove_rare_stopword)

# Cek hasil 
df[["stopword_removal", 'stopword_removal_clean']].head()

# Hanya ambil kata dari kolom stopword_removal
all_words_stopword = [word for tokens in df["stemming_output"] for word in tokens]

# Hitung frekuensi spesifik hanya untuk kolom stopword_removal
word_counts_stopword = Counter(all_words_stopword)

# Tentukan rare words berdasarkan kolom ini saja
rare_threshold = 5
rare_words_stopword = {word for word, count in word_counts_stopword.items() if count < rare_threshold}

# Bersihkan
def remove_rare_stopword(tokens):
    return [word for word in tokens if word not in rare_words_stopword]

df['stemming_output_clean'] = df['stemming_output'].apply(remove_rare_stopword)

# Cek hasil 
df[["stemming_output", 'stemming_output_clean']].head()

### ✅ **9. Delete Empty Rows**

# Cek jumlah baris yang memiliki token kosong setelah preprocessing
df_empty = df[df["stemming_output_clean"].apply(lambda x: len(x) == 0)]
print(f"Jumlah baris kosong: {len(df_empty)}")

### ✅ **Gabungkan Kembali Token**

# Gabungkan list token menjadi satu string
df['konten_bersih_stemming'] = df['stemming_output_clean'].apply(lambda tokens: ' '.join(tokens))
df['konten_bersih_tanpa_stemming'] = df['stopword_removal_clean'].apply(lambda tokens: ' '.join(tokens))

Pemisahan dua kolom hasil preprocessing dilakukan untuk membandingkan dua pendekatan berbeda dalam pengolahan teks, yaitu menggunakan stemming (**final_stemming_output_clean**) dan tanpa stemming (**final_stopword_removal_clean**). 
Kolom dengan stemming mengubah kata ke bentuk dasarnya untuk mengurangi variasi kata dan menyederhanakan fitur, sementara kolom tanpa stemming mempertahankan bentuk asli kata setelah menghapus stopword. Perbedaan ini memungkinkan evaluasi performa model berdasarkan teknik preprocessing yang digunakan.

## 📌 **Save Data**

# Simpan DataFrame
output_path = "../data/hasil_preprocessing.xlsx"
df.to_excel(output_path, index=False)
